refactor(client): extract field tests to sibling file
